Use simple pagination for survey progress
Showing the question numbers is difficult and bug-ridden because
it uses the session to store the comment count, and the length
of the survey can change based on skip logic.

It's much more straightforward if we just show the page number
which is good enough to give the user an indication of progress.