Add rules to segment admin
wagtail_personalisation uses `AbstractBaseRule.__subclasses__()` to
discover which rules to show in the interface which is incredibly
confusing.

It's not strictly necessary to inline the rules in the admin, they
only needed to be imported here so that Python is aware that they exist.
However I think it's more explicit and is what wagtail_personalisation
does in `admin.py` (even though that only contains a subset of their
rules, the others are loaded because they are subclasses of
AbstractBaseRule.

I'm unable to write a meaningful test for this because for some reason
the test is already loading `gem.rules`, so the test passes even without
this code change.