Create custom template tag to create id string from checkbox field label
`field.id_for_label` elvaluates to None in the form template. This means that checkboxes ended up with the same id, resulting in incorrect behaviour when checking boxes.

This is a temporary solution and warrants further investigation as to why this particular method fails.

The reason for the custom template, is so that the `<input>` is alongside the `<label>` element, rather than nested inside it, so that it can be styled appropriately.