Refactor BBM logic into context processor
Each of these `request.get_host` in the template is a call to the
`get_host()` method on a `WSGIRequest`. It's hard to tell because
in the template syntax we can call a method without the parentheses.

This probably isn't a massive performance problem but it's unnecessary.
It's better to group this logic together so that we only run it once
in the context processor. This also means we can test it easily.